requirements.txt
streamlit>=1.36
ultralytics>=8.2
opencv-python>=4.9
pillow>=10.0
pandas>=2.0
pywebview>=5.0
fastapi>=0.111
uvicorn[standard]>=0.30
python-multipart>=0.0.9
